requirements.txt
langchain
langchain-community
langchain_groq
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain_huggingface
langchain-chroma
faiss-cpu